updated mock model generator
